pyproject.toml
[project]
name = "agentify-design2code"
version = "0.1.0"
description = "Agentified Design2Code - Assessment framework for HTML generation from screenshots."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "litellm>=1.0.0",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "opencv-python>=4.8.0",
    "beautifulsoup4>=4.12.0",
    "colormath>=3.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ftfy>=6.1.0",
    "regex>=2023.0.0",
    "open-clip-torch>=2.20.0",
    "playwright>=1.40.0",
    "google-adk[a2a]>=0.1.0",
]
